LPS-86239 dropIsValid function moved to LayoutDropUtils
