+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/miller_columns/utils/LayoutUtils.es.js
@@ -1,5 +1,4 @@
 import {setIn} from '../../utils/utils.es';
-import {DROP_TARGET_TYPES} from './LayoutDragDrop.es';
 
 /**
  * Append an item to a column and returns a new array of columns
@@ -115,39 +114,6 @@ function deleteEmptyColumns(layoutColumns) {
 	}
 
 	return nextLayoutColumns;
-}
-
-/**
- * @param {object} sourceItem
- * @param {string} sourceItemColumnIndex
- * @param {string} targetId
- * @param {string} targetType
- * @return {boolean} Returns whether a drop is valid or not
- * @review
- */
-function dropIsValid(
-	sourceItem,
-	sourceItemColumnIndex,
-	targetId,
-	targetType
-) {
-	let targetColumnIsChild = false;
-	let targetEqualsSource = false;
-
-	if (targetType === DROP_TARGET_TYPES.column) {
-		targetColumnIsChild = columnIsItemChild(
-			targetId,
-			sourceItem,
-			sourceItemColumnIndex
-		);
-	}
-	else if (targetType === DROP_TARGET_TYPES.item) {
-		targetEqualsSource = (sourceItem.plid === targetId);
-	}
-
-	const targetExists = (targetId !== null);
-
-	return targetExists && !targetEqualsSource && !targetColumnIsChild;
 }
 
 /**
@@ -441,7 +407,6 @@ export {
 	clearPath,
 	columnIsItemChild,
 	deleteEmptyColumns,
-	dropIsValid,
 	getColumnActiveItem,
 	getColumnLastItem,
 	getItem,
